# docs/swagger.yaml
definitions:
  api.ApiError:
    properties:
      error:
        type: string
    type: object
...
  types.PaidRequest:
    properties:
      amount:
        type: integer
      operation:
        type: string
      order_id:
        type: string
      payment_id:
        type: string
    type: object
  types.PaymentRequest:
    properties:
      amount:
        type: integer
      card_expiry_month:
        type: string
      card_expiry_year:
        type: string
      card_number:
        type: string
      card_security_code:
        type: string
      currency:
        type: string
      id:
        type: string
      order_id:
        type: string
    type: object
  types.PaymentResponse:
    properties:
      id:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: Simple payment system
  title: Payment Application
  version: "1.0"
paths:
...
  /payment/auth:
    post:
      consumes:
      - application/json
      description: 'Create payment: Acceptance of payment'
      parameters:
      - description: create payment info
        in: path
        name: id
        required: true
        type: string
      - description: create payment info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/types.PaymentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.PaymentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ApiError'
      summary: Create payment
      tags:
      - Payment
  /payment/cancel/{id}:
    post:
      consumes:
      - application/json
      description: 'Cancel payment: cancel authorization payment'
      parameters:
      - description: cancel payment info
        in: path
        name: id
        required: true
        type: string
      - description: cancel payment info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/types.PaidRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.PaymentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ApiError'
      summary: Cancel payment
      tags:
      - Payment
  /payment/capture/{id}:
    post:
      consumes:
      - application/json
      description: 'Capture payment: Successful payment'
      parameters:
      - description: capture payment info
        in: path
        name: id
        required: true
        type: string
      - description: capture payment info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/types.PaidRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.PaymentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ApiError'
      summary: Capture payment
      tags:
      - Payment
  /payment/refund/{id}:
    post:
      consumes:
      - application/json
      description: 'Refund: Refunded payment, if there is a refund'
      parameters:
      - description: refund payment info
        in: path
        name: id
        required: true
        type: string
      - description: refund payment info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/types.PaidRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.PaymentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ApiError'
      summary: Refund payment
      tags:
      - Payment
securityDefinitions:
  "":
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"